Nabil: changements
le tri variable fonctionne

il fallait changer l'ordre des indice de les tableau presenceVariable de toutes les contraintes

il fallait aussi changer le type de l'ID en float

+++ Donnees/backtrack.c
@@ -121,7 +121,7 @@ void forward_check(variables X_instanciee){
 	while( !fin_de_liste(V)){
 		l = V->contraintes;
 		while(l!= NULL){
-			if((l->presenceVariable)[X_instanciee->id]=='1') {
+			if((l->presenceVariable)[(int)X_instanciee->id]=='1') {
 				while(affecter_valeur(V)){
 					if (evalue_contrainte(l->a))
 						ajouter_valeur(&D,V->valeur);
